fix/refactor : [#5] 사용하지 않는 코드 정리, 통합 테스트 오류 수정
