+++ src/main/java/hhplus/lecture/application/service/LectureService.java
@@ -1,13 +1,13 @@
 package hhplus.lecture.application.service;
 
-import hhplus.lecture.domain.model.Lecture;
-import hhplus.lecture.infrastructure.persistence.LectureItemEntity;
+import hhplus.lecture.domain.LectureItemEntity;
 import hhplus.lecture.infrastructure.repository.InstructorRepository;
 import hhplus.lecture.infrastructure.repository.LectureItemRepository;
 import hhplus.lecture.infrastructure.repository.LectureRepository;
-import hhplus.lecture.infrastructure.persistence.LectureEntity;
+import hhplus.lecture.domain.LectureEntity;
 import hhplus.lecture.interfaces.dto.lecture.LectureDto;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -15,6 +15,7 @@ import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional
 public class LectureService {
 
     private final LectureRepository lectureRepository;
@@ -22,13 +23,13 @@ public class LectureService {
     private final InstructorRepository instructorRepository;
 
     // 특강 정보 조회
-    public Lecture getLecture(String lectureCode) {
+    public LectureEntity getLecture(String lectureCode) {
         LectureEntity entity = lectureRepository.findByLectureCode(lectureCode);
         if(entity == null) {
             throw new NoSuchElementException("특강을 찾을 수 없습니다.");
         }
 
-        return entity.toDomain();
+        return entity;
     }
 
     // 모든 특강 조회
