+++ src/main/java/hhplus/lecture/infrastructure/repository/InstructorRepository.java
@@ -1,6 +1,6 @@
 package hhplus.lecture.infrastructure.repository;
 
-import hhplus.lecture.infrastructure.persistence.InstructorEntity;
+import hhplus.lecture.domain.InstructorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstructorRepository extends JpaRepository<InstructorEntity, Long> {
